requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.2
fastapi==0.112.2
uvicorn==0.30.6
